Began reworking SerializedTransferProtocol
Adjusted some global documentation and introduced _minimum_packet_size private variable and minimum_payload_size constructor argument to allow enforcing a particular number of bytes to be present in the reception buffer before triggering reception procedure.  This optimization is adapted from the python version of the class (now deprecated) and may be helpful for some platforms. Mostly this is needed for platforms that use OSes like Windows, which has a very large overhead for 'read' operations, so there is a high cost of attempting to read data and being certain that the operation succeeds is important to avoid wasting CPU time.