Moved minimum payload size from constructor to template arguments
The way this parameter is used aligns more with template parameter logic rather than the constructor argument logic and also merits from having a static assertion test that requires compile-time constancy enforced via template specification. Specifically, in the previous version, when the argument was set to an invalid value it was implicitly rectified, which is not the best way of handling erroneous values of this parameter given its overall importance./ In the current version the user is instead notified that the parameter value is invalid during compilation and is prompted to fix the error.